5758 - Add validation repository folders
